Consolidate X platform adapter, client, and schema into single file
